pyproject.toml
[tool.poetry]
name = "pubmed-fetcher"
version = "0.1.0"
description = "A tool to fetch PubMed papers with non-academic authors"
authors = ["Mayuri <mayuri.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.79"
requests = "^2.26.0"
pandas = "^1.3.0"
click = "^8.0.1"

[tool.poetry.scripts]
get-papers-list = "scripts.get_papers_list:get_papers_list"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"